Extract print functions to print.rs (refs #86)
Moves print_response(), print_hierarchy(), print_workspace_list(), and
print_pane_list() out of main.rs into a dedicated print.rs module.
Pure formatting functions with no shared state.

First extraction in amux-cli decomposition.

No behavior change — pure mechanical refactor.
